refactor(ai): complete V1 following system removal and V2 consolidation
Major architectural cleanup:
- Remove old V1 following system completely (opponentBlocking, teammateSupport, etc.)
- Consolidate followingV2/ into following/ directory
- Eliminate feature switching between V1/V2 systems
- Update AI_SYSTEM.md with split leading/following decision diagrams
- Fix import paths and type errors after restructuring
- Clean up unused test imports and incomplete tests
- Implement pair preservation in strategic selection utilities

Key changes:
- Unified AI system now uses only V2 following strategy
- Enhanced documentation with separate leading/following flowcharts
- Preserved multiComboFollowingStrategy as essential component
- Fixed all TypeScript compilation and ESLint issues
