+++ __tests__/ai/aiGameContext.test.ts
@@ -499,7 +499,7 @@ describe("AI Game Context", () => {
 
       expect(result.isTeammateWinning).toBe(true);
       expect(result.isOpponentWinning).toBe(false);
-      expect(result.isSelfWinning).toBe(false);
+      // Note: isSelfWinning property removed - test needs update
       expect(result.currentWinner).toBe(PlayerId.Bot3);
     });
 
@@ -520,7 +520,7 @@ describe("AI Game Context", () => {
 
       expect(result.isTeammateWinning).toBe(false);
       expect(result.isOpponentWinning).toBe(true);
-      expect(result.isSelfWinning).toBe(false);
+      // Note: isSelfWinning property removed - test needs update
       expect(result.currentWinner).toBe(PlayerId.Human);
     });
 
@@ -541,7 +541,7 @@ describe("AI Game Context", () => {
 
       expect(result.isTeammateWinning).toBe(false);
       expect(result.isOpponentWinning).toBe(false);
-      expect(result.isSelfWinning).toBe(true);
+      // Note: isSelfWinning property removed - test needs update
       expect(result.currentWinner).toBe(PlayerId.Bot1);
     });
 
@@ -581,7 +581,9 @@ describe("AI Game Context", () => {
 
       const result = analyzeTrickWinner(gameState, PlayerId.Bot1);
 
-      expect(result.canBeatCurrentWinner).toBeDefined();
+      // Basic assertions for trick winner analysis
+      expect(result.currentWinner).toBe(PlayerId.Human);
+      expect(result.trickPoints).toBe(0);
     });
   });
