+++ __tests__/helpers/ai.ts
@@ -48,10 +48,8 @@ export const createAIGameContext = (
     currentWinner: PlayerId.Human, // Default to human for test scenarios
     isTeammateWinning: false,
     isOpponentWinning: false,
-    isSelfWinning: false,
     isLeadWinning: false,
     trickPoints: 0,
-    canBeatCurrentWinner: false,
     isTrumpLead: false,
     isCurrentlyTrumped: false,
   },
