+++ src/ai/aiStrategy.ts
@@ -2,7 +2,6 @@ import { Card, Combo, GameState, Player, PositionStrategy } from "../types";
 import { gameLogger } from "../utils/gameLogger";
 import { analyzeCombo, createGameContext } from "./aiGameContext";
 import { selectFollowingPlay } from "./following/followingStrategy";
-import { selectFollowingPlay as selectFollowingPlayV2 } from "./followingV2/followingStrategy";
 import { selectLeadingPlay } from "./leading/leadingStrategy";
 
 /**
@@ -50,32 +49,15 @@ export function makeAIPlay(
       analysis: analyzeCombo(combo, trumpInfo, context),
     }));
 
-    // Feature toggle for enhanced following algorithm
-    const USE_ENHANCED_FOLLOWING = true; // Toggle to switch between V1 and V2
-
-    let restructuredPlay: Card[];
-
-    if (USE_ENHANCED_FOLLOWING) {
-      // V2: Enhanced following algorithm (no fallback - V2 handles its own errors)
-      restructuredPlay = selectFollowingPlayV2(
-        comboAnalyses,
-        context,
-        {} as PositionStrategy,
-        trumpInfo,
-        gameState,
-        player.id,
-      );
-    } else {
-      // V1: Original following algorithm
-      restructuredPlay = selectFollowingPlay(
-        comboAnalyses,
-        context,
-        {} as PositionStrategy,
-        trumpInfo,
-        gameState,
-        player.id,
-      );
-    }
+    // Use enhanced following algorithm V2
+    const restructuredPlay = selectFollowingPlay(
+      comboAnalyses,
+      context,
+      {} as PositionStrategy,
+      trumpInfo,
+      gameState,
+      player.id,
+    );
     gameLogger.debug("ai_following_decision", {
       decisionPoint: "following_play",
       player: player.id,
