+++ src/ai/following/crossSuitDecision.ts
