+++ src/ai/following/sameSuitDecision.ts
