+++ src/ai/following/validCombosDecision.ts
