+++ src/ai/following/voidDecision.ts
